[StateMachine] refactor to move classes to where they actually belong. Basic StateMachine handling is done in Resource/ResourceBundle. PaymentStates and Transitions are inside Payment/PaymentBundle. OrderStates and basic Order Workflow is done in Order/OrderBundle. The rest is done in CoreBundle
