+++ src/CoreShop/Bundle/CoreBundle/Order/StateResolver/OrderStateResolver.php
@@ -10,15 +10,17 @@
  * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
  */
 
-namespace CoreShop\Bundle\CoreBundle\StateMachine\StateResolver;
+namespace CoreShop\Bundle\CoreBundle\Order\StateResolver;
 
-use CoreShop\Bundle\CoreBundle\StateMachine\OrderPaymentStates;
-use CoreShop\Bundle\CoreBundle\StateMachine\OrderShippingStates;
-use CoreShop\Bundle\CoreBundle\StateMachine\OrderTransitions;
-use CoreShop\Bundle\CoreBundle\StateMachine\StateMachineManager;
-use CoreShop\Component\Core\Model\OrderInterface;
+use CoreShop\Component\Core\OrderPaymentStates;
+use CoreShop\Component\Order\Model\OrderInterface;
+use CoreShop\Component\Order\OrderShippingStates;
+use CoreShop\Component\Order\OrderTransitions;
+use CoreShop\Component\Order\StateResolver\StateResolverInterface;
+use CoreShop\Component\Resource\Workflow\StateMachineManager;
+use Webmozart\Assert\Assert;
 
-final class OrderStateResolver implements StateMachineResolverInterface
+final class OrderStateResolver implements StateResolverInterface
 {
     /**
      * @var StateMachineManager
@@ -51,6 +53,11 @@ final class OrderStateResolver implements StateMachineResolverInterface
      */
     private function canOrderBeComplete(OrderInterface $order)
     {
+        /**
+         * @var $order \CoreShop\Component\Core\Model\OrderInterface
+         */
+        Assert::isInstanceOf($order, \CoreShop\Component\Core\Model\OrderInterface::class);
+
         return
             OrderPaymentStates::STATE_PAID === $order->getPaymentState() &&
             OrderShippingStates::STATE_SHIPPED === $order->getShippingState();
