+++ src/CoreShop/Bundle/CoreBundle/Resources/config/app/workflow/coreshop_payment.yml
@@ -1,0 +1,14 @@
+core_shop_resource:
+  state_machine:
+    callbacks:
+      coreshop_payment:
+        after:
+          # implement processors?
+          #process_order:
+            #on: ['fail', 'cancel']
+            #do: ['@coreshop.order_payment_processor', 'process']
+            #args: ['object.getOrder()']
+          resolve_state:
+            on: ['complete', 'refund']
+            do: ['@coreshop.state_machine_resolver.order_payment', 'resolve']
+            args: ['object.getOrder()']
