+++ src/CoreShop/Bundle/CoreBundle/Resources/config/services.yml
@@ -18,11 +18,11 @@ imports:
   - { resource: "services/cart-processor.yml" }
   - { resource: "services/reports.yml" }
   - { resource: "services/pimcore_installer.yml" }
-  - { resource: "services/state_machine.yml" }
   - { resource: "services/inventory.yml" }
   - { resource: "services/handler.yml" }
   - { resource: "services/commands.yml" }
   - { resource: "services/locale.yml" }
+  - { resource: "services/workflow.yml" }
 
 services:
   coreshop.core.key_transformer:
