+++ src/CoreShop/Bundle/OrderBundle/Resources/config/app/config.yml
@@ -1,3 +1,6 @@
+imports:
+    - { resource: "workflow.yml" }
+
 jms_serializer:
     metadata:
         directories:
