+++ src/CoreShop/Bundle/OrderBundle/Resources/config/services/workflow.yml
@@ -1,46 +1,10 @@
-parameters:
-  coreshop.workflow.manager.order.state_identifier: 'Order State Change'
-
 services:
-  coreshop.workflow.manager.registry:
-    class: CoreShop\Component\Registry\ServiceRegistry
-    arguments:
-      - 'CoreShop\Component\Order\Workflow\WorkflowManagerInterface'
-      - 'workflow-managers'
+    coreshop.state_machine.marking_store.single_state.order_state:
+        class: Symfony\Component\Workflow\MarkingStore\SingleStateMarkingStore
+        arguments:
+            - orderState
 
-  coreshop.workflow.manager.order.validators:
-    class: CoreShop\Component\Registry\PrioritizedServiceRegistry
-    arguments:
-      - 'CoreShop\Component\Order\Workflow\ProposalValidatorInterface'
-      - 'workflow-order-manager-validators'
-
-  coreshop.workflow.manager.order:
-    class: CoreShop\Bundle\OrderBundle\Workflow\WorkflowManager
-    arguments:
-      - '@coreshop.workflow.manager.order.validators'
-      - '@event_dispatcher'
-      - 'CoreShop\Component\Order\Model\OrderInterface'
-      - '@pimcore.translator'
-      - '%coreshop.workflow.manager.order.state_identifier%'
-      - '@security.token_storage'
-    tags:
-      - {name: coreshop.workflow.manager, type: '%coreshop.model.order.class%' }
-
-  coreshop.workflow.manager.order.validator.payment:
-    class: CoreShop\Bundle\OrderBundle\Workflow\Order\PaymentValidator
-    tags:
-        - {name: coreshop.workflow.validator, type: payment, priority: 10, manager: coreshop.workflow.manager.order}
-
-  coreshop.workflow.manager.order.validator.invoice:
-    class: CoreShop\Bundle\OrderBundle\Workflow\Order\OrderDocumentValidator
-    arguments:
-      - '@coreshop.order.invoice.processable'
-    tags:
-        - {name: coreshop.workflow.validator, type: invoice, priority: 20, manager: coreshop.workflow.manager.order}
-
-  coreshop.workflow.manager.order.validator.shipment:
-    class: CoreShop\Bundle\OrderBundle\Workflow\Order\OrderDocumentValidator
-    arguments:
-      - '@coreshop.order.shipment.processable'
-    tags:
-        - {name: coreshop.workflow.validator, type: shipment, priority: 30, manager: coreshop.workflow.manager.order}
+    coreshop.state_machine.marking_store.order_pimcore_marketing_store:
+        class: CoreShop\Component\Resource\Workflow\MarketingStore\PimcorePersistentMarkingStore
+        arguments:
+          - '@coreshop.state_machine.marking_store.single_state.order_state'
