+++ src/CoreShop/Bundle/OrderBundle/Resources/install/pimcore/classes/CoreShopOrder.json
@@ -2995,6 +2995,97 @@
                         "locked": false
                     },
                     {
+                        "fieldtype": "panel",
+                        "labelWidth": 100,
+                        "layout": null,
+                        "name": "Status",
+                        "type": null,
+                        "region": null,
+                        "title": "Status",
+                        "width": null,
+                        "height": null,
+                        "collapsible": false,
+                        "collapsed": false,
+                        "bodyStyle": "",
+                        "datatype": "layout",
+                        "permissions": null,
+                        "childs": [
+                            {
+                                "fieldtype": "input",
+                                "width": null,
+                                "queryColumnType": "varchar",
+                                "columnType": "varchar",
+                                "columnLength": 190,
+                                "phpdocType": "string",
+                                "regex": "",
+                                "unique": false,
+                                "name": "orderState",
+                                "title": "Order State",
+                                "tooltip": "",
+                                "mandatory": false,
+                                "noteditable": true,
+                                "index": false,
+                                "locked": false,
+                                "style": "",
+                                "permissions": null,
+                                "datatype": "data",
+                                "relationType": false,
+                                "invisible": false,
+                                "visibleGridView": false,
+                                "visibleSearch": false
+                            },
+                            {
+                                "fieldtype": "input",
+                                "width": null,
+                                "queryColumnType": "varchar",
+                                "columnType": "varchar",
+                                "columnLength": 190,
+                                "phpdocType": "string",
+                                "regex": "",
+                                "unique": false,
+                                "name": "shippingState",
+                                "title": "Shipping State",
+                                "tooltip": "",
+                                "mandatory": false,
+                                "noteditable": true,
+                                "index": false,
+                                "locked": false,
+                                "style": "",
+                                "permissions": null,
+                                "datatype": "data",
+                                "relationType": false,
+                                "invisible": false,
+                                "visibleGridView": false,
+                                "visibleSearch": false
+                            },
+                            {
+                                "fieldtype": "input",
+                                "width": null,
+                                "queryColumnType": "varchar",
+                                "columnType": "varchar",
+                                "columnLength": 190,
+                                "phpdocType": "string",
+                                "regex": "",
+                                "unique": false,
+                                "name": "invoiceState",
+                                "title": "Invoice State",
+                                "tooltip": "",
+                                "mandatory": false,
+                                "noteditable": true,
+                                "index": false,
+                                "locked": false,
+                                "style": "",
+                                "permissions": null,
+                                "datatype": "data",
+                                "relationType": false,
+                                "invisible": false,
+                                "visibleGridView": false,
+                                "visibleSearch": false
+                            }
+                        ],
+                        "locked": false
+                    },
+                    {
                         "fieldtype": "tabpanel",
                         "name": "Numbers",
                         "type": null,
