+++ src/CoreShop/Bundle/PaymentBundle/Resources/config/services.yml
@@ -1,2 +1,3 @@
 imports:
     - { resource: "services/form.yml" }
+    - { resource: "services/workflow.yml" }
