+++ src/CoreShop/Bundle/PaymentBundle/Resources/config/services/workflow.yml
@@ -1,0 +1,11 @@
+services:
+    coreshop.state_machine.marking_store.single_state.state:
+        class: Symfony\Component\Workflow\MarkingStore\SingleStateMarkingStore
+        arguments:
+            - state
+
+    coreshop.state_machine.marking_store.payment_orm_marketing_store:
+        class: CoreShop\Component\Resource\Workflow\MarketingStore\OrmPersistentMarkingStore
+        arguments:
+            - '@coreshop.state_machine.marking_store.single_state.state'
+            - '@doctrine'
