+++ src/CoreShop/Bundle/PayumBundle/Extension/UpdatePaymentStateExtension.php
@@ -2,10 +2,10 @@
 
 namespace CoreShop\Bundle\PayumBundle\Extension;
 
-use CoreShop\Bundle\CoreBundle\StateMachine\PaymentTransitions;
-use CoreShop\Bundle\CoreBundle\StateMachine\StateMachineManager;
 use CoreShop\Bundle\PayumBundle\Request\GetStatus;
 use CoreShop\Component\Payment\Model\PaymentInterface;
+use CoreShop\Component\Payment\PaymentTransitions;
+use CoreShop\Component\Resource\Workflow\StateMachineManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Payum\Core\Extension\Context;
 use Payum\Core\Extension\ExtensionInterface;
