+++ src/CoreShop/Bundle/ResourceBundle/DependencyInjection/Configuration.php
@@ -16,6 +16,7 @@ use CoreShop\Bundle\ResourceBundle\Controller\ResourceController;
 use CoreShop\Bundle\ResourceBundle\CoreShopResourceBundle;
 use CoreShop\Component\Resource\Factory\Factory;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
+use Symfony\Component\Config\Definition\Builder\NodeBuilder;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
@@ -33,6 +34,7 @@ final class Configuration implements ConfigurationInterface
         $this->addTranslationsSection($rootNode);
         $this->addDriversSection($rootNode);
         $this->addPimcoreResourcesSection($rootNode);
+        $this->addWorkflowSection($rootNode);
 
         return $treeBuilder;
     }
@@ -154,5 +156,49 @@ final class Configuration implements ConfigurationInterface
                 ->end()
             ->end()
         ->end();
+    }
+
+
+    /**
+     * @param ArrayNodeDefinition $node
+     */
+    private function addWorkflowSection(ArrayNodeDefinition $node)
+    {
+        $callbacks = $node
+            ->children()
+                ->arrayNode('state_machine')
+                    ->children()
+                        ->arrayNode('callbacks')
+                            ->useAttributeAsKey('name')
+                            ->prototype('array')
+                            ->children();
+
+        $this->addSubCallbackSection($callbacks, 'guard');
+        $this->addSubCallbackSection($callbacks, 'before');
+        $this->addSubCallbackSection($callbacks, 'after');
+
+        $callbacks->end()->end()->end()->end()->end();
+
+    }
+
+    /**
+     * @param NodeBuilder $callbacks
+     * @param string      $type
+     */
+    protected function addSubCallbackSection(NodeBuilder $callbacks, $type)
+    {
+        $callbacks
+            ->arrayNode($type)
+                ->useAttributeAsKey('name')
+                ->prototype('array')
+                    ->children()
+                        ->variableNode('on')->end()
+                        ->variableNode('do')->end()
+                        ->scalarNode('priority')->defaultValue(0)->end()
+                        ->arrayNode('args')->performNoDeepMerging()->prototype('scalar')->end()
+                    ->end()
+                ->end()
+            ->end()
+        ;
     }
 }
