+++ src/CoreShop/Bundle/ResourceBundle/EventListener/Workflow/WorkflowListener.php
@@ -10,10 +10,8 @@
  * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
  */
 
-namespace CoreShop\Bundle\CoreBundle\StateMachine\EventListener;
+namespace CoreShop\Bundle\ResourceBundle\EventListener\Workflow;
 
-use CoreShop\Component\Pimcore\VersionHelper;
-use Pimcore\Model\DataObject\Concrete;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
