+++ src/CoreShop/Bundle/ResourceBundle/Resources/config/services/workflow.yml
@@ -1,0 +1,25 @@
+services:
+    coreshop.state_machine_manager:
+        class: CoreShop\Component\Resource\StateMachine\StateMachineManager
+        arguments:
+            - '@workflow.registry'
+
+    coreshop.state_machine_applier:
+        class: CoreShop\Component\Resource\StateMachine\StateMachineApplier
+        arguments:
+            - '@coreshop.state_machine_manager'
+
+    coreshop.state_machine_cascade_transition:
+        public: true
+        class: CoreShop\Component\Resource\Workflow\Callback\CascadeTransition
+        arguments:
+            - '@coreshop.state_machine_manager'
+
+    ## event handler
+    coreshop.state_machine_listener_processor:
+        class: CoreShop\Bundle\ResourceBundle\EventListener\Workflow\WorkflowListener
+        arguments:
+            - '%coreshop.state_machine.callbacks%'
+            - '@service_container'
+        tags:
+            - { name: kernel.event_subscriber }
