+++ src/CoreShop/Component/Core/Model/OrderInterface.php
@@ -20,40 +20,10 @@ interface OrderInterface extends BaseOrderInterface, CarrierAwareInterface
     /**
      * @return string
      */
-    public function getOrderState();
-
-    /**
-     * @param string $orderState
-     */
-    public function setOrderState($orderState);
-
-    /**
-     * @return string
-     */
     public function getPaymentState();
 
     /**
      * @param string $paymentState
      */
     public function setPaymentState($paymentState);
-
-    /**
-     * @return string
-     */
-    public function getShippingState();
-
-    /**
-     * @param string $shippingState
-     */
-    public function setShippingState($shippingState);
-
-    /**
-     * @return string
-     */
-    public function getInvoiceState();
-
-    /**
-     * @param string $invoiceState
-     */
-    public function setInvoiceState($invoiceState);
 }
