+++ src/CoreShop/Component/Order/Model/OrderInterface.php
@@ -20,6 +20,36 @@ interface OrderInterface extends SaleInterface
     /**
      * @return string
      */
+    public function getOrderState();
+
+    /**
+     * @param string $orderState
+     */
+    public function setOrderState($orderState);
+
+    /**
+     * @return string
+     */
+    public function getShippingState();
+
+    /**
+     * @param string $shippingState
+     */
+    public function setShippingState($shippingState);
+
+    /**
+     * @return string
+     */
+    public function getInvoiceState();
+
+    /**
+     * @param string $invoiceState
+     */
+    public function setInvoiceState($invoiceState);
+
+    /**
+     * @return string
+     */
     public function getOrderLanguage();
 
     /**
