+++ src/CoreShop/Component/Order/OrderShippingStates.php
@@ -10,19 +10,14 @@
  * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
  */
 
-declare(strict_types=1);
+namespace CoreShop\Component\Order;
 
-namespace CoreShop\Bundle\CoreBundle\StateMachine;
-
-final class PaymentTransitions
+final class OrderShippingStates
 {
-    const IDENTIFIER = 'coreshop_payment';
-
-    const TRANSITION_CREATE = 'create';
-    const TRANSITION_PROCESS = 'process';
-    const TRANSITION_COMPLETE = 'complete';
-    const TRANSITION_FAIL = 'fail';
-    const TRANSITION_CANCEL = 'cancel';
-    const TRANSITION_REFUND = 'refund';
-    const TRANSITION_VOID = 'void';
+    const STATE_INITIALIZED = 'initialized';
+    const STATE_NEW = 'new';
+    const STATE_READY = 'ready';
+    const STATE_CANCELLED = 'cancelled';
+    const STATE_PARTIALLY_SHIPPED = 'partially_shipped';
+    const STATE_SHIPPED = 'shipped';
 }
