+++ src/CoreShop/Component/Order/OrderStates.php
@@ -10,13 +10,12 @@
  * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
  */
 
-namespace CoreShop\Bundle\CoreBundle\StateMachine;
+namespace CoreShop\Component\Order;
 
-final class OrderTransitions
+final class OrderStates
 {
-    const IDENTIFIER = 'coreshop_order';
-
-    const TRANSITION_CREATE = 'create';
-    const TRANSITION_CANCEL = 'cancel';
-    const TRANSITION_COMPLETE = 'complete';
+    const STATE_INITIALIZED = 'initialized';
+    const STATE_NEW = 'new';
+    const STATE_CANCELLED = 'cancelled';
+    const STATE_COMPLETE = 'complete';
 }
